requirements.txt
trfl
pyyaml==5.4.1
protobuf==3.20.*
gym
pandas
matplotlib
numpy==1.23.1
numba
ipykernel